Don't parse date strings in get_object_storage_network_statistics
This changes the `get_object_storage_network_statistics` method to not accept arbitrary strings that might or might not contain dates.

The parsing of those strings expects them to be in one strict format and in the local time zone, which may or may not be the case.
Instead, it's simpler to just only accept `datetime.datetime` objects and call `.isoformat()` on them.

This has the happy consequence of being able to remove the `dateutil` dependency too.